requirements.txt
alembic==1.8.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
Deprecated==1.2.13
fastapi==0.85.1
fastapi-jwt-auth==0.5.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
Mako==1.2.3
MarkupSafe==2.1.1
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pluggy==1.0.0
psycopg2==2.9.4
py==1.11.0
pydantic==1.10.2
PyJWT==1.7.1
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
validators==0.20.0
wrapt==1.14.1
